Fix: generalise bind_tbl()
bind_tbl.default resulted in a stop(), and only bind_tbl.user_list was useful. However, there may be cases when a user_list is manipualted and loses it's class. So, instread, the functionality was moved into bind_tbl.default with a warning when an object of the wrong class is used.